# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fire Prevention API",
        "description": "API documentation with JWT authentication",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/readings": {
            "get": {
                "tags": [
                    "Sensor Readings"
                ],
                "summary": "Get all sensor readings",
                "description": "Returns a list of all sensor readings with their associated device information",
                "operationId": "e06a175f29cd83ec2d2dd23c0d1faa89",
                "responses": {
                    "200": {
                        "description": "List of sensor readings retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SensorReadingResponse"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No sensor readings found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No sensor readings found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sensor Readings"
                ],
                "summary": "Create a new sensor reading",
                "description": "Creates a new sensor reading with the provided measurements",
                "operationId": "98de13709ed8eb63c07d1d828935d8a5",
                "requestBody": {
                    "description": "Sensor reading data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "device_token",
                                    "temperature",
                                    "humidity",
                                    "smoke_level",
                                    "gas_level"
                                ],
                                "properties": {
                                    "device_token": {
                                        "description": "Token of the device making the reading",
                                        "type": "string",
                                        "example": "abc123"
                                    },
                                    "temperature": {
                                        "description": "Temperature reading in Celsius",
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.5
                                    },
                                    "humidity": {
                                        "description": "Humidity percentage",
                                        "type": "number",
                                        "format": "float",
                                        "example": 60
                                    },
                                    "smoke_level": {
                                        "description": "Smoke level reading",
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "gas_level": {
                                        "description": "Gas level reading",
                                        "type": "number",
                                        "format": "float",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sensor reading created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sensor reading created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "device_token": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The device token is required"
                                            ]
                                        },
                                        "temperature": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The temperature must be a number"
                                            ]
                                        },
                                        "humidity": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The humidity must be a number"
                                            ]
                                        },
                                        "smoke_level": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The smoke level must be a number"
                                            ]
                                        },
                                        "gas_level": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The gas level must be a number"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SensorReadingResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "device_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "temperature": {
                        "type": "number",
                        "format": "float",
                        "example": 25.5
                    },
                    "humidity": {
                        "type": "number",
                        "format": "float",
                        "example": 60
                    },
                    "smoke_level": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "gas_level": {
                        "type": "number",
                        "format": "float",
                        "example": 15
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-01-29T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-01-29T12:00:00Z"
                    },
                    "device": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Sensor Cocina"
                            },
                            "location": {
                                "type": "string",
                                "example": "Cocina Principal"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}